+++ docs/versionhistory.rst
@@ -12,6 +12,9 @@ This library adheres to `Semantic Versioning 2.0 <https://semver.org/#semantic-v
 - Added support for ``LiteralString``
 - Added support for ``TypeGuard``
 - Added support for the subclassable ``Any`` on Python 3.11 and ``typing_extensions``
+- Added configuration option ``collection_check_strategy`` and changed the default to
+  only check the first item of collections (list, set, abstractset, dict, sequence,
+  mapping)
 - ``check_return_value`` now uses ``TypeGuard`` in its own return type annotation,
   making it more useful when used together with static type checkers
 - Made it possible to apply ``@typechecked`` on top of ``@classmethod`` /